chore(restructuring components): Putting Drag N Drop components to its own directory
