+++ client/components/DragAndDropComponents/DraggableMultipleChoice/DraggableMultipleChoice.tsx
@@ -1,16 +1,16 @@
-import React from "react";
+import React, { useState } from "react";
 import { Controller, useFieldArray } from "react-hook-form";
 import { Draggable } from "react-beautiful-dnd";
-import { DraggableDropdownProps } from "./types";
-import { useEffect, useState } from "react";
-import styles from "./DraggableDropdown.module.css";
-import MinusCircleIcon from "../../assets/MinusCircleIcon";
-import PlusCircleIcon from "../../assets/PlusCircleIcon";
+import { DraggableMultipleChoiceProps } from "./types";
+import styles from "./DraggableMultipleChoice.module.css";
+import PlusCircleIcon from "../../../assets/PlusCircleIcon";
+import MinusCircleIcon from "../../../assets/MinusCircleIcon";
+import EllipseIcon from "../../../assets/EllipseIcon";
 import DraggableSubMenu from "../DraggableSubMenu/DraggableSubMenu";
 import Switch from "react-switch";
 import { useTranslation } from "react-i18next";
 
-const DraggableDropdown: React.FC<DraggableDropdownProps> = ({
+const DraggableMultipleChoice: React.FC<DraggableMultipleChoiceProps> = ({
   id,
   index,
   title,
@@ -31,27 +31,12 @@ const DraggableDropdown: React.FC<DraggableDropdownProps> = ({
     name: `fields.${index}.properties.choices`,
   });
 
-  const [options, setOptions] = useState(fields);
-
-  useEffect(() => {
-    setOptions(fields);
-  }, [fields]);
-
   const addOption = () => {
     append({ ref: `option-${fields.length + 1}`, label: "" });
   };
 
   const removeOption = (optionIndex: number) => {
     remove(optionIndex);
-  };
-
-  const changeOptionsIndex = (label: string, optionIndex: number) => {
-    const updatedOptions = [...options];
-    updatedOptions[optionIndex] = {
-      ...updatedOptions[optionIndex],
-      label,
-    };
-    setOptions(updatedOptions);
   };
 
   return (
@@ -64,7 +49,7 @@ const DraggableDropdown: React.FC<DraggableDropdownProps> = ({
           style={provided.draggableProps.style}
           onClick={handleClick}>
           <div style={{ position: "relative" }}>
-            <p className={styles.textElementTypeText}>{t("dropDown")}</p>
+            <p className={styles.textElementTypeText}>{t("multipleChoice")}</p>
             <div
               style={{
                 padding: 16,
@@ -99,10 +84,9 @@ const DraggableDropdown: React.FC<DraggableDropdownProps> = ({
                 </div>
               )}
               <Controller
-                key={index}
                 name={`fields.${index}.title`}
                 control={control}
-                defaultValue={title}
+                defaultValue={title} // Ensure the default value is set
                 render={({ field }) => (
                   <>
                     <input
@@ -123,6 +107,12 @@ const DraggableDropdown: React.FC<DraggableDropdownProps> = ({
                     gap: 10,
                     marginTop: 10,
                   }}>
+                  <EllipseIcon
+                    width={15}
+                    height={15}
+                    color={"#AAAAAA"}
+                    fill={"white"}
+                  />
                   <Controller
                     key={index}
                     name={`fields.${index}.properties.choices.${idx}.label`}
@@ -130,10 +120,6 @@ const DraggableDropdown: React.FC<DraggableDropdownProps> = ({
                     render={({ field }) => (
                       <input
                         {...field}
-                        onChange={(e) => {
-                          field.onChange(e); // Ensure the field value is updated in react-hook-form
-                          changeOptionsIndex(e.target.value, idx); // Update the selectOptions state
-                        }}
                         placeholder={`Option ${idx + 1}`}
                         style={{
                           marginLeft: -5,
@@ -144,10 +130,7 @@ const DraggableDropdown: React.FC<DraggableDropdownProps> = ({
                       />
                     )}
                   />
-                  <button
-                    type="button"
-                    onClick={() => removeOption(idx)}
-                    style={{ marginLeft: "8px" }}>
+                  <button type="button" onClick={() => removeOption(idx)}>
                     <MinusCircleIcon width={20} height={20} color={"#FF0000"} />
                   </button>
                 </div>
@@ -176,4 +159,4 @@ const DraggableDropdown: React.FC<DraggableDropdownProps> = ({
   );
 };
 
-export default DraggableDropdown;
+export default DraggableMultipleChoice;
